translation_service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-generativeai